Don't overwrite existing routes
'Adding' an existing route would overwrite with the default template.

+++ src/sanic_kit/console.py
@@ -339,15 +339,16 @@ class SanicKit(App):
         new_dir = Path("src/routes") / route
         new_dir.mkdir(exist_ok=True, parents=True)
         new_page = new_dir / "+page.sanic"
-        new_page.touch()
-        new_page.write_text(
-            """\
+        if not new_page.exists():
+            new_page.touch()
+            new_page.write_text(
+                """\
 <handler>
 </handler>
 {% block main %}
 {% endblock %}
 """
-        )
+            )
         await self.query_one(Routes).refresh_tree(new_page)
 
     def compose(self):
